+++ uw-httpclient/src/main/java/uw/httpclient/json/JsonObjectMapperImpl.java
@@ -9,7 +9,7 @@ import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
-import uw.common.util.DateUtils;
+import uw.common.util.DateTools;
 import uw.httpclient.exception.DataMapperException;
 import uw.httpclient.http.DataObjectMapper;
 
@@ -144,13 +144,13 @@ public class JsonObjectMapperImpl implements DataObjectMapper {
             @Override
             public Date deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException, JacksonException {
                 String dateString = jsonParser.getText();
-                return DateUtils.stringToDate(dateString);
+                return DateTools.stringToDate(dateString);
             }
         });
         dateUtilModule.addSerializer(Date.class, new JsonSerializer<Date>() {
             @Override
             public void serialize(Date date, JsonGenerator jsonGenerator, SerializerProvider serializerProvider) throws IOException, JacksonException {
-                jsonGenerator.writeString(DateUtils.dateToString(date, DateUtils.DATE_MILLIS_ISO));
+                jsonGenerator.writeString(DateTools.dateToString(date, DateTools.DATE_MILLIS_ISO));
             }
         });
         jsonMapper.registerModule(dateUtilModule);